requirements.txt
attrs==20.2.0             # via pytest
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
django==3.1.2
idna==2.10                 # via requests
importlib-metadata==2.0.0  # via pluggy, pytest
more-itertools==8.5.0     # via pytest
packaging==20.4           # via pytest
pillow==7.2.0
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==3.10.0
pytest==6.1.1             # via pytest-django
pytz==2020.1              # via django
requests==2.24.0
six==1.15.0               # via packaging
sorl-thumbnail==12.6.3
sqlparse==0.4.1           # via django
urllib3==1.25.10           # via requests
wcwidth==0.2.5            # via pytest
zipp==3.3.0               # via importlib-metadata
django-debug-toolbar
djangorestframework
markdown
django-filter

